[Prover Service] Rename smoke.rs -> prover_handler.rs
