+++ prover-service/src/tests/mod.rs
@@ -1,8 +1,8 @@
 // Copyright (c) Aptos Foundation
 pub mod federated_jwk;
 pub mod jwk_fetcher;
+pub mod prover_handler;
 pub mod request_handler;
-pub mod smoke;
 pub mod training_wheels;
 pub mod types;
 pub mod utils;
